Move session tests to their own file
